Move command helper function; Make error message a little less wordy
